Support for new asyncio coroutines
The new asyncio API is very powerful, and much easier than working with selectors directly.

This patch adds the coroutines `async_read` and `async_read_one`, and makes `read_loop()` compatible with the new `async for` syntax.

Unfortunately, the code turned out a bit ugly in order to make the syntax backward compatible with Python 2.7.