Added IOCTL to retrieve the destination uniq as well as phys if available, added uniq as a property to InputDevice
Needed because some later kernels (at least from 4.15) have started reporting the bluetooth controller
MAC for bluetooth devices as the phys, rather than the actual physical address of the device (which is
obviously the bit you actually care about in most cases). As with phys, there's no guarantee that uniq
is actually set.