FIX: ensure lists handed in to _uinput.create are long enough
This prevents a

E       SystemError: Objects/longobject.c:415: bad argument to internal function

when passing in too short of a list on newer versions of cpython.

The underlying issue is that in uinput.c:uinput_create the list access
is not checked for IndexError.  If out of range, it returns null and
sets an error.  At some point PyLong_AsLong started to fail when
passed null (rather than I assume treating it as 0).

This ensures filling with 0 behavior at the python level.