Split "create" and "setup" parts for uinput devices creation
In order to be able to enable EV_FF in uinput device, ff_max_effects
fields needs to be set to a non-zero value. However the ordering of
execution in uinput.py does not allow for that because indvidual
events are enabled before _uinput.create() is invoked.

In order to accomodate EV_FF, split _uinput.create() into two
functions:
    - _uinput.setup(), which does the majority of device
    configuration.

    - _uinput.create(), which only creates the device without
      changing any of its configuration

On top of that, change the code in uinput.py to make the call to
setup() before individual events are enabled to enable adding
ff_max_effects configuration in follow-up commits.