uinput: avoid possible unterminated string.
This is a false positive because the kernel does the right thing and takes
UINPUT_MAX_NAME_SIZE into account. But this shuts up coverity and other static
analyzers so I think we can live with a 79 character restriction instead of
80.

Also change to use sizeof(), not that it really makes a difference but it
allows for the array size to be changed while leaving the constant as-is,
impossible as that is.