Wait 0.1s if /dev/input/event* does not exist
Unlike what a previous commit claimed, the new method is only guaranteed
to find the device name instantaneously, but if /dev is managed by udev,
the corresponding /dev/input/input* file might not exist immediately. On
modern Linux systems, those files are managed by devtmpfs so they do
show up immediately and it would be a waste of time to sleep, but we do
want a fallback time.sleep() call for those operating systems which do
not use devtmpfs.

It is assumed that on any linux system, /sys is managed by sysfs; it
conceptually doesn't really make sense to let that directory be managed
by any other filesystem because the files in /sys are basically
syscalls. As such, we assume that /sys always updates instantaneously,
and if for whatever reason it doesn't, _find_device_fallback shall be
used instead.