New method for finding the device node corresponding to a uinput device (#206)
* Add binding for UI_GET_SYSNAME ioctl

* New modern Linux-specific method to find device

The old method of scanning the filesystem to find an event device with a
matching name was prone to race conditions, required a time.sleep delay,
unnecessarily opened many event devices, and had other fixable issues.

The new UI_GET_SYSNAME-based method is immune to race conditions and
does not require time.sleep(), while also not suffering from the
aforementioned fixable issues.

* Move get_sysname() to the same block as _find_device_linux()

Co-authored-by: Tobi <28510156+sezanzeb@users.noreply.github.com>

* Sort event devices in _find_device_fallback()

* Wait 0.1s if /dev/input/event* does not exist

Unlike what a previous commit claimed, the new method is only guaranteed
to find the device name instantaneously, but if /dev is managed by udev,
the corresponding /dev/input/input* file might not exist immediately. On
modern Linux systems, those files are managed by devtmpfs so they do
show up immediately and it would be a waste of time to sleep, but we do
want a fallback time.sleep() call for those operating systems which do
not use devtmpfs.

It is assumed that on any linux system, /sys is managed by sysfs; it
conceptually doesn't really make sense to let that directory be managed
by any other filesystem because the files in /sys are basically
syscalls. As such, we assume that /sys always updates instantaneously,
and if for whatever reason it doesn't, _find_device_fallback shall be
used instead.

* Add a comment

Co-authored-by: Tobi <28510156+sezanzeb@users.noreply.github.com>

* Replace IOError with OSError

Since Python 3.3 (PEP 3151), IOError and OSError are aliases of each
other, and OSError is the prefered name. Rename all mentions of IOError
in the C bindings with OSError.

* Replace IOError with OSError in documentation

* Explain why the sorting is necessary

Co-authored-by: Tobi <28510156+sezanzeb@users.noreply.github.com>

---------

Co-authored-by: Tobi <28510156+sezanzeb@users.noreply.github.com>