Wait up to two seconds for input device. (#215)
It has been reported that even on modern Linux distributions that use
devtmpfs, it can take longer than 0.1 seconds before the input device
becomes available with the right permissions. If sysfs reports that a
specific input device should be used, but it cannot be opened
immediately, keep trying again for up to two seconds.